refactor_and_assets: change to use Engine event based Input system
